refactoring :: ApplicationsDomainInterface Transfer 나눔
