+++ Projects/Domain/ApplicationsDomain/Interface/DTO/DataTransfer/TotalPassStudentTransfer.swift
@@ -1,0 +1,11 @@
+import Foundation
+
+public extension TotalPassStudentResponseDTO {
+    func toDomain() -> TotalPassStudentEntity {
+        TotalPassStudentEntity(
+            totalStudentCount: totalStudentCount,
+            passedCount: passedCount,
+            approvedCount: approvedCount
+        )
+    }
+}
